Pipe updated
