+++ .gitignore
@@ -116,6 +116,7 @@ venv.bak/
 # Other
 *.sql
 *.txt
+*.old
 
 # Other Folders
 sandbox/
